refactor: move methods to BaseClient
